+++ src/CronSchedulerWorker/TestJob.cs
@@ -10,15 +10,17 @@ using Microsoft.Extensions.Options;
 
 namespace CronSchedulerWorker
 {
-    public class TestJob : ScheduledJob
+    public class TestJob : IScheduledJob
     {
         private readonly TestJobOptions _options;
         private readonly ILogger<TestJob> _logger;
 
-        public TestJob(IOptionsMonitor<TestJobOptions> options, ILogger<TestJob> logger, IHostLifetime lifetime)
-            : base(options.CurrentValue)
+        public TestJob(
+            IOptionsMonitor<TestJobOptions> options,
+            ILogger<TestJob> logger,
+            IHostLifetime lifetime)
         {
-            _options = options.CurrentValue;
+            _options = options.Get(Name);
 
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
 
@@ -26,7 +28,9 @@ namespace CronSchedulerWorker
             _logger.LogInformation("IHostLifetime: {hostLifetime}", lifetime.GetType());
         }
 
-        public override async Task ExecuteAsync(CancellationToken cancellationToken)
+        public string Name { get; } = nameof(TestJob);
+
+        public async Task ExecuteAsync(CancellationToken cancellationToken)
         {
             _logger.LogInformation("{jobName} is executing with Custom Field Value: {customField}", nameof(TestJob), _options.CustomField);